First round of suggestions from code review. Initial moves, renames and deletions
